add app icon and minor roomdb changes

+++ app/src/main/java/com/example/openrestaurant/activity/ShowFavouritesActivity.kt
@@ -1,11 +1,12 @@
 package com.example.openrestaurant.activity
 
-import androidx.appcompat.app.AppCompatActivity
+import android.content.DialogInterface
 import android.os.Bundle
 import android.view.View
 import android.widget.ProgressBar
 import android.widget.TextView
 import android.widget.Toast
+import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
@@ -14,13 +15,20 @@ import com.example.openrestaurant.adapter.FavouriteItemOnClicked
 import com.example.openrestaurant.adapter.FavouriteOrderAdapter
 import com.example.openrestaurant.roomdb.Favourite
 import com.example.openrestaurant.roomdb.FavouriteViewModel
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
+import com.google.firebase.firestore.ktx.firestore
+import com.google.firebase.ktx.Firebase
+import io.paperdb.Paper
 
 class ShowFavouritesActivity : AppCompatActivity(), FavouriteItemOnClicked {
     private lateinit var favouriteViewModel: FavouriteViewModel
     private lateinit var recyclerView: RecyclerView
     private lateinit var favouriteAdapter: FavouriteOrderAdapter
+    private val db = Firebase.firestore
+    private lateinit var restaurantId: String
 
     override fun onCreate(savedInstanceState: Bundle?) {
+        Paper.init(this)
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_show_favourites)
         supportActionBar?.elevation = 0f
@@ -30,6 +38,7 @@ class ShowFavouritesActivity : AppCompatActivity(), FavouriteItemOnClicked {
         recyclerView = findViewById(R.id.favouriteRecyclerView)
         findViewById<TextView>(R.id.favouriteEmptyMessage).visibility = View.GONE
         findViewById<ProgressBar>(R.id.favouriteProgressBar).visibility = View.VISIBLE
+        restaurantId = Paper.book().read("RESTAURANT_ID")
 
         recyclerView.apply {
             layoutManager = LinearLayoutManager(this@ShowFavouritesActivity)
@@ -43,7 +52,7 @@ class ShowFavouritesActivity : AppCompatActivity(), FavouriteItemOnClicked {
         favouriteViewModel.allFavourites.observe(this, {
             it?.let {
                 findViewById<ProgressBar>(R.id.favouriteProgressBar).visibility = View.GONE
-                favouriteAdapter.updateFavouriteOrders(it)
+                favouriteAdapter.updateFavouriteOrders(it.reversed())
                 if (it.isEmpty()) {
                     findViewById<TextView>(R.id.favouriteEmptyMessage).visibility = View.VISIBLE
                 }
@@ -52,11 +61,22 @@ class ShowFavouritesActivity : AppCompatActivity(), FavouriteItemOnClicked {
     }
 
     override fun onClicked(favourite: Favourite) {
-        Toast.makeText(this, "Favourite Tapped!", Toast.LENGTH_SHORT).show()
+        Toast.makeText(this,
+            "${favourite.order_id}, ${favourite.restaurant_id}",
+            Toast.LENGTH_SHORT).show()
     }
 
     override fun onDelete(favourite: Favourite) {
-        favouriteViewModel.deleteFavourite(favourite)
-        Toast.makeText(this, "Favourite Deleted!", Toast.LENGTH_SHORT).show()
+        MaterialAlertDialogBuilder(this)
+            .setTitle("Delete Favourite Order?")
+            .setPositiveButton("Delete") { dialogInterface, _ ->
+                favouriteViewModel.deleteFavourite(favourite)
+                Toast.makeText(this, "Favourite Deleted!", Toast.LENGTH_SHORT).show()
+                dialogInterface.dismiss()
+            }
+            .setNegativeButton("Cancel") { dialogInterface, _ ->
+                dialogInterface.cancel()
+            }
+            .show()
     }
 }
